refactor(habitat): move hook operations into policy
