+++ docs/projects/habitat-harness/deep-refactor/service-shape-backlog.md
@@ -138,3 +138,4 @@ Completed burn-downs:
 - Structural check execution lanes now have named command/Nx-backed and file-layer/staged-path policies; `execution.policy.ts` is reduced to rule selection and lane orchestration instead of owning every execution implementation.
 - Hook procedure context vocabulary now lives in a named module-local policy file; `hook/module.ts` no longer exports raw provider/port/state type declarations as inline implementation clutter.
 - Boundary taxonomy filesystem/config/Nx graph acquisition now lives under `src/validation`; shared `service/model/graph` retains the pure taxonomy parser/auditor instead of importing filesystem and Nx graph APIs.
+- Hook pre-commit/pre-push operation bodies now live in module-local `procedure-operations.policy.ts`; `hook/module.ts` is reduced to effect-oRPC context projection plus shared check-report context construction.
